Dodaj magućnost da korisnik izbriše svoj recept
Drugi obični korisnici naravno ne mogu izbrisati tuđe recepte.

+++ IzvorniKod/Undercooked-Demo/src/main/java/hr/fer/progi/UndercookedDemo/rest/RecipeController.java
@@ -3,6 +3,7 @@ package hr.fer.progi.UndercookedDemo.rest;
 import hr.fer.progi.UndercookedDemo.domain.Recipe;
 import hr.fer.progi.UndercookedDemo.service.PersonService;
 import hr.fer.progi.UndercookedDemo.service.RecipeService;
+import hr.fer.progi.UndercookedDemo.service.RequestDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -40,6 +41,18 @@ public class RecipeController {
 	@DeleteMapping("/admin/{id}")
 	@PreAuthorize("hasAuthority('SCOPE_ROLE_ADMIN')")
 	public void adminDelete(@PathVariable("id") Long id) {
+		recipeService.deleteById(id);
+	}
+
+	@DeleteMapping("/{id}")
+	@PreAuthorize("hasAuthority('SCOPE_ROLE_USER')")
+	public void userDelete(@PathVariable("id") Long id, Principal principal) {
+		var person = personService.fromPrincipal(principal);
+		var existing = recipeService.findById(id);
+
+		if(!existing.getAuthor().id().equals(person.getId()))
+			throw new RequestDeniedException("Tried to delete another person's recipe.");
+
 		recipeService.deleteById(id);
 	}
 }
